- move template examples inside src folder
