+++ cli-typescript/src/templates/example/erc1155_example.json
