+++ cli-typescript/src/templates/example/erc721_example.json
