+++ cli-typescript/src/templates/index.ts
